Feat/os rm

+++ P1/src/structs/indice.h
@@ -14,7 +14,7 @@ typedef struct indice
 
 // FUNCTIONS
 
-Indice *indice_init(unsigned long int tamano, unsigned int identificador_relativo, unsigned int identificador_absoluto);
+Indice *indice_init(unsigned int tamano, unsigned int identificador_relativo, unsigned int identificador_absoluto);
 
 void assing_data_list(Indice *indice, Data *data, int index);
